central_brain/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.2
langgraph>=0.2.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
